server/requirements.txt
google-adk
pydantic
litellm